+++ src/cpu/utils/utils.hh
@@ -14,11 +14,11 @@ namespace cpu::utils
     using vector_t = Matrix::vector_t;
     using matrix_t = Matrix::matrix_t;
 
-    double compute_distance(const vector_t& p, const vector_t& q);
-    void get_nearest_neighbors(const matrix_t& P, const matrix_t& Q, matrix_t& res, std::vector<double>& distances);
+    float compute_distance(const vector_t& p, const vector_t& q);
+    void get_nearest_neighbors(const matrix_t& P, const matrix_t& Q, matrix_t& res, std::vector<float>& distances);
     unsigned int get_line_count(const std::string& path);
 
-    void save_result(std::size_t iteration, double error);
+    void save_result(std::size_t iteration, float error);
 
     void string_split(std::string str, const std::string& delimiter, std::vector<std::string>& words);
 } // namespace utils
